Start the feature
https://coveord.atlassian.net/browse/JSUI-2674
