+++ src/rest/Facet/FacetRequest.ts
@@ -17,7 +17,11 @@ export enum FacetType {
   /**
    * Request facet values representing ranges of dates.
    */
-  dateRange = 'dateRange'
+  dateRange = 'dateRange',
+  /**
+   * Request facet values representing a hierarchy.
+   */
+  hierarchical = 'hierarchical'
 }
 
 /**
@@ -152,4 +156,17 @@ export interface IFacetRequest {
    * **Default (Search API):** `false`
    */
   generateAutomaticRanges?: boolean;
+
+  /**
+   * The character to use to split field values into a hierarchical sequence.
+   *
+   * **Example:**
+   * For a multi-value field containing the following values: `c; c&gt;folder2; c&gt;folder2&gt;folder3;`
+   * The delimiting character is `&gt;`.
+   * For a hierarchical field containing the following values: `c;folder2;folder3;`
+   * The delimiting character is `;`.
+   * 
+   * **Default (Search API):** `;`
+   */
+  delimitingCharacter?: string;
 }
